% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posts.R
\name{get_reddit_posts}
\alias{get_reddit_posts}
\title{Retrieve post data from a subreddit}
\usage{
get_reddit_posts(subreddit, token, agent, username, limit = 100,
  max_pages = 10)
}
\arguments{
\item{subreddit}{name of subreddit}

\item{token}{oauth token obtained using \code{\link{get_reddit_token}}}

\item{agent}{name of your reddit api application}

\item{username}{your reddit username}

\item{limit}{the number of posts to return per page (100 maximum)}

\item{max_pages}{the maximum number of pages return}
}
\value{
dataframe containing data for posts in the subreddit
}
\description{
Uses the OAuth API URL https://oauth.reddit.com/r/subreddit/new.json
}
