% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{get_reddit_token}
\alias{get_reddit_token}
\title{Retrieve reddit API access token}
\usage{
get_reddit_token(agent, client_id, secret, username, password)
}
\arguments{
\item{agent}{name of your reddit api application}

\item{client_id}{client_id of your reddit api application}

\item{secret}{secret of your reddit api application}

\item{username}{your reddit username}

\item{password}{your reddit password}
}
\value{
oauth token string
}
\description{
Retrieve reddit API access token
}
